requirements.txt
# Core Dependencies
pandas>=1.5.0
numpy>=1.24.0
requests>=2.28.0
python-dotenv>=1.0.0
pydantic>=2.0.0
structlog>=23.0.0

# API and Data Processing
kafka-python>=2.0.2
pyspark>=3.4.0
pyarrow>=12.0.0
boto3>=1.26.0
psycopg2-binary>=2.9.0

# Data Transformation
dbt-core>=1.6.0
dbt-postgres>=1.6.0
dbt-utils>=1.1.0

# Visualization
streamlit>=1.25.0
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Monitoring and Logging
prometheus-client>=0.17.0
grafana-api>=1.0.3

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Jupyter and Analysis
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=6.5.0

# Additional Utilities
schedule>=1.2.0
python-dateutil>=2.8.0
pytz>=2023.3